Relocate ember-array internals
