+++ packages/@ember/array/index.ts
@@ -3,7 +3,6 @@
 */
 import { DEBUG } from '@glimmer/env';
 import { PROXY_CONTENT } from '@ember/-internals/metal';
-import { isEmberArray, setEmberArray } from '@ember/-internals/utils';
 import {
   objectAt,
   replaceInNativeArray,
@@ -22,6 +21,7 @@ import { ENV } from '@ember/-internals/environment';
 import Observable from '@ember/object/observable';
 import type { MethodNamesOf, MethodParams, MethodReturns } from '@ember/-internals/utils/types';
 import type { ComputedPropertyCallback } from '@ember/-internals/metal';
+import { isEmberArray, setEmberArray } from '@ember/array/-internals';
 
 export { makeArray } from '@ember/-internals/utils';
