+++ types/publish.mjs
@@ -161,7 +161,6 @@ const PREVIEW_MODULES = [
   '@ember/-internals/utils/index.d.ts',
   '@ember/-internals/utils/lib/cache.d.ts',
   '@ember/-internals/utils/lib/dictionary.d.ts',
-  '@ember/-internals/utils/lib/ember-array.d.ts',
   '@ember/-internals/utils/lib/get-debug-name.d.ts',
   '@ember/-internals/utils/lib/guid.d.ts',
   '@ember/-internals/utils/lib/inspect.d.ts',
@@ -203,6 +202,7 @@ const PREVIEW_MODULES = [
   '@ember/array/index.d.ts',
   '@ember/array/mutable.d.ts',
   '@ember/array/proxy.d.ts',
+  '@ember/array/-internals.d.ts',
   '@ember/canary-features/index.d.ts',
   '@ember/component/helper.d.ts',
   '@ember/component/index.d.ts',
